Release: v0.0.5
This release of the `bloombox-elements` suite of elements presents an
all-new development framework for Polymer, and many bugfixes and
improvements. In particular, Bloombox completed it's `private-beta`
phase on `v0.0.4`, and much of the feedback received from Abatin is
incorporated in this release.

This is also the first release where each element is itself fully
released in some form, with at least a tagged version of `0.0.1`.

The following components have been added this round:
- `<bloombox-account>`: Displays account information for a partner-level
  account, such as payment info and EIN.

Additionally, support for the following has been added to each element:
- Makefile and gulpfile improvements
- CodeClimate static code analysis
- Internal Buildbot and Artifactory